defer: Move basic inlineCallbacks tests to test_inlinecb
deferredGenerator and inlineCallbacks tests just happen to be similar
enough that some of them can be shared. However, this sharing is
ephemeral because there's no underlying reason why both sets of tests
should change both at the same time. Changes to how inlineCallbacks
works will affect only that part of the test suite and changes to how
deferredGenerator will affect only its part of the test suite as well.

Duplicating the tests makes them easier to discover and understand.
Additionally, this will need to be done at some point anyway because
deferredGenerator has been deprecated.
